Refactor DatabaseLink into it's own package
