+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/JoinEvent.java
@@ -2,6 +2,7 @@ package com.github.gcc_minecraft_team.sps_mc_link_spigot;
 
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.claims.ClaimHandler;
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.claims.Team;
+import database.DatabaseLink;
 import net.md_5.bungee.api.ChatMessageType;
 import net.md_5.bungee.api.chat.ClickEvent;
 import net.md_5.bungee.api.chat.ComponentBuilder;
