+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/LeaveEvent.java
@@ -1,5 +1,6 @@
 package com.github.gcc_minecraft_team.sps_mc_link_spigot;
 
+import database.DatabaseLink;
 import org.bukkit.ChatColor;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
