+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/claims/ClaimCommands.java
@@ -1,6 +1,6 @@
 package com.github.gcc_minecraft_team.sps_mc_link_spigot.claims;
 
-import com.github.gcc_minecraft_team.sps_mc_link_spigot.DatabaseLink;
+import database.DatabaseLink;
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.SPSSpigot;
 import fr.mrmicky.fastboard.FastBoard;
 import org.bukkit.ChatColor;
