+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/claims/Team.java
@@ -1,6 +1,6 @@
 package com.github.gcc_minecraft_team.sps_mc_link_spigot.claims;
 
-import com.github.gcc_minecraft_team.sps_mc_link_spigot.DatabaseLink;
+import database.DatabaseLink;
 import org.bson.codecs.pojo.annotations.BsonProperty;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.configuration.serialization.ConfigurationSerializable;
