+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/discord/DiscordCommands.java
@@ -1,6 +1,6 @@
 package com.github.gcc_minecraft_team.sps_mc_link_spigot.discord;
 
-import com.github.gcc_minecraft_team.sps_mc_link_spigot.DatabaseLink;
+import database.DatabaseLink;
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.PluginConfig;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
