+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/moderation/ModerationCommands.java
@@ -1,6 +1,6 @@
 package com.github.gcc_minecraft_team.sps_mc_link_spigot.moderation;
 
-import com.github.gcc_minecraft_team.sps_mc_link_spigot.DatabaseLink;
+import database.DatabaseLink;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
