+++ src/com/github/gcc_minecraft_team/sps_mc_link_spigot/permissions/PermissionsHandler.java
@@ -1,6 +1,6 @@
 package com.github.gcc_minecraft_team.sps_mc_link_spigot.permissions;
 
-import com.github.gcc_minecraft_team.sps_mc_link_spigot.DatabaseLink;
+import database.DatabaseLink;
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.SPSSpigot;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.MemorySection;
