+++ src/database/DatabaseLink.java
@@ -1,5 +1,6 @@
-package com.github.gcc_minecraft_team.sps_mc_link_spigot;
+package database;
 
+import com.github.gcc_minecraft_team.sps_mc_link_spigot.SPSSpigot;
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.claims.ClaimHandler;
 import com.github.gcc_minecraft_team.sps_mc_link_spigot.claims.Team;
 import com.mongodb.*;
